+++ TestConvexHull.cpp
@@ -1,7 +1,7 @@
 #include "TestConvexHull.h"
 #include "SFML/Graphics.hpp"
 
-TestConvexHull::TestConvexHull(const std::vector<glm::vec2>& _points,  const std::string& _outputFile) : ConvexHullBase(_points, _outputFile)
+TestConvexHull::TestConvexHull(const std::vector<glm::ivec2>& _points,  const std::string& _outputFile) : ConvexHullBase(_points, _outputFile)
 {
 }
 
@@ -12,7 +12,7 @@ TestConvexHull::~TestConvexHull()
 void TestConvexHull::Draw(sf::RenderWindow& window)
 {
 	sf::VertexArray vertices;
-	for(glm::vec2 p : m_points)
+	for(glm::ivec2 p : m_points)
 	{
 		sf::Vertex point(sf::Vector2f(p.x, p.y), sf::Color::White);
 		vertices.append(point);
@@ -20,7 +20,7 @@ void TestConvexHull::Draw(sf::RenderWindow& window)
 	window.draw(vertices);
 }
 
-std::vector<glm::vec2> TestConvexHull::GenerateConvexHull()
+std::vector<glm::ivec2> TestConvexHull::GenerateConvexHull()
 {
 	//Should be saving the convex hull
 	SavePointsToFile(m_points);
