+++ TestConvexHull.h
@@ -4,11 +4,11 @@
 class TestConvexHull : public ConvexHullBase
 {
 public:
-	TestConvexHull(const std::vector<glm::vec2>& _points, const std::string& _outputFile = "");
+	TestConvexHull(const std::vector<glm::ivec2>& _points, const std::string& _outputFile = "");
 	~TestConvexHull();
 
 	void Draw(sf::RenderWindow& window) override;
-	std::vector<glm::vec2> GenerateConvexHull() override;
+	std::vector<glm::ivec2> GenerateConvexHull() override;
 private:
 
 };
